ARTEMIS-5845 ignore divert bindings w/removing address
When removing an address, the broker checks to see if there are any
bindings so as not to orphan any queues. However, a divert is also a
binding so the presence of a divert will prevent the deletion of an
address as well.

This is particularly problematic for retroactive addresses because they
will always have a divert. This means that the broker can't delete a
retroactive address automatically.

+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/postoffice/impl/PostOfficeImpl.java
@@ -883,18 +883,19 @@ public class PostOfficeImpl implements PostOffice, NotificationListener, Binding
             server.callBrokerAddressPlugins(plugin -> plugin.beforeRemoveAddress(address));
          }
 
-         final Collection<Binding> bindingsForAddress = getDirectBindings(address);
-         if (force) {
-            for (Binding binding : bindingsForAddress) {
-               if (binding instanceof LocalQueueBinding localQueueBinding) {
-                  localQueueBinding.getQueue().deleteQueue(true);
-               } else if (binding instanceof RemoteQueueBinding) {
-                  removeBinding(binding.getUniqueName(), null, true);
+         if (isAddressBound(address)) {
+            if (force) {
+               for (Binding binding : getDirectBindings(address)) {
+                  if (binding instanceof LocalQueueBinding localQueueBinding) {
+                     localQueueBinding.getQueue().deleteQueue(true);
+                  } else if (binding instanceof RemoteQueueBinding) {
+                     removeBinding(binding.getUniqueName(), null, true);
+                  }
                }
-            }
 
-         } else if (!bindingsForAddress.isEmpty()) {
-            throw ActiveMQMessageBundle.BUNDLE.addressHasBindings(address);
+            } else {
+               throw ActiveMQMessageBundle.BUNDLE.addressHasBindings(address);
+            }
          }
          managementService.unregisterAddress(address);
          final AddressInfo addressInfo = addressManager.removeAddressInfo(address);
@@ -1063,7 +1064,7 @@ public class PostOfficeImpl implements PostOffice, NotificationListener, Binding
 
    @Override
    public boolean isAddressBound(final SimpleString address) throws Exception {
-      Collection<Binding> bindings = getDirectBindings(address);
+      Collection<Binding> bindings = getDirectBindings(address).stream().filter(binding -> binding.getType() != BindingType.DIVERT).toList();
       return bindings != null && !bindings.isEmpty();
    }
